+++ example/lib/src/app/view/my_app.dart
@@ -18,6 +18,7 @@ class MyApp extends AppStatefulWidgetMVC {
 
 ///
 class _MyAppState extends AppStateMVC<MyApp> {
+  //
   factory _MyAppState() => _this ??= _MyAppState._();
 
   _MyAppState._()
@@ -35,30 +36,10 @@ class _MyAppState extends AppStateMVC<MyApp> {
         );
   static _MyAppState? _this;
 
-  /// Optionally you can is the framework's buildApp() function
-  /// instead of its build() function and allows for the InheritWidget feature
+  /// Used this function instead of its build() function
+  /// and allows for the InheritWidget feature
   @override
   Widget buildApp(BuildContext context) => MaterialApp(
-        home: FutureBuilder<bool>(
-            future: initAsync(),
-            builder: (context, snapshot) {
-              //
-              if (snapshot.hasData) {
-                //
-                if (snapshot.data!) {
-                  /// Key identifies the widget. New key? New widget!
-                  /// Demonstrates how to explicitly 're-create' a State object
-                  return Page1(key: UniqueKey());
-                } else {
-                  //
-                  return const Text('Failed to startup');
-                }
-              } else if (snapshot.hasError) {
-                //
-                return Text('${snapshot.error}');
-              }
-              // By default, show a loading spinner.
-              return const Center(child: CircularProgressIndicator());
-            }),
+        home: Page1(key: UniqueKey()),
       );
 }