+++ example/lib/src/home/view/page_02.dart
@@ -26,7 +26,7 @@ class Page2 extends StatefulWidget {
 class _Page2State extends InheritedStateMVC<Page2, _Page02Inherited> {
   /// Define an InheritedWidget to be inserted above this Widget on the Widget tree.
   _Page2State()
-      : super(inheritedWidget: (child) => _Page02Inherited(child: child));
+      : super(inheritedBuilder: (child) => _Page02Inherited(child: child));
 
   @override
   void initState() {